+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - TT-1964 - Update save for later page & email
 - TT-1957 - Collapse open commodity nodes
 - TT-1975 - rate limiting integration sender ip_address
+- TT-1918 - Set ZenDesk-Servicename
 
 ### Bug fixes